improve performance & upgrade interface
